Time: 0 ms (100.00%) | Memory: 19.7 MB (44.63%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -19,7 +19,7 @@ public:
 
         while (!q.empty()){
             int size = q.size();
-            int minIdx = q.front().second;
+            ll minIdx = q.front().second;
             ll first = -1, last = -1;
 
             for (int i = 0; i < size; i++){
